Expanded ResourceRegistry.Delete to include deletion of related policy (#642)
Expanded ResourceRegistry.Delete to include deletion of related policy file to prevent deleting of resources leading to orphan policy files.
Aslo added a new PolicyRepository.TryDeletePolicyAsync that deletes all snapsnots and versions to ensure proper cleanup.

+++ src/Altinn.ResourceRegistry.Core/Services/ResourceRegistryService.cs
@@ -79,6 +79,12 @@ namespace Altinn.ResourceRegistry.Core.Services
         /// <inheritdoc/>
         public async Task Delete(string id, CancellationToken cancellationToken = default)
         {
+            bool deletePolicyResult = await _policyRepository.TryDeletePolicyAsync(id, cancellationToken);
+            if (!deletePolicyResult)
+            {
+                throw new AccessManagementUpdateException($"Deleting policy for resource {id} failed");
+            }
+
             await _repository.DeleteResource(id, cancellationToken);
         }
